move user consents test into a separate test suite for DIA-2940-consent-applies-fix
