+++ samples/app/src/androidTest/java/com/sourcepoint/app/v6/MainActivityKotlinTest.kt
@@ -4,7 +4,6 @@ import android.preference.PreferenceManager
 import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.launchActivity
 import androidx.test.internal.runner.junit4.AndroidJUnit4ClassRunner
-import androidx.test.platform.app.InstrumentationRegistry
 import com.example.uitestutil.*
 import com.sourcepoint.app.v6.TestUseCase.Companion.checkAllCcpaConsentsOn
 import com.sourcepoint.app.v6.TestUseCase.Companion.checkSomeConsentsOff
@@ -49,7 +48,6 @@ import com.sourcepoint.cmplibrary.exception.CampaignType
 import com.sourcepoint.cmplibrary.model.MessageLanguage
 import com.sourcepoint.cmplibrary.model.exposed.CcpaStatus
 import com.sourcepoint.cmplibrary.model.exposed.SpCampaign
-import com.sourcepoint.cmplibrary.util.userConsents
 import io.mockk.mockk
 import io.mockk.verify
 import kotlinx.coroutines.delay
@@ -63,8 +61,6 @@ import java.util.UUID
 
 @RunWith(AndroidJUnit4ClassRunner::class)
 class MainActivityKotlinTest {
-
-    private val appContext by lazy { InstrumentationRegistry.getInstrumentation().targetContext }
 
     lateinit var scenario: ActivityScenario<MainActivityKotlin>
 
@@ -1156,27 +1152,6 @@ class MainActivityKotlinTest {
             PreferenceManager.getDefaultSharedPreferences(activity).run {
                 getInt("sp.key.config.propertyId", 0).assertEquals(newPropertyId)
             }
-        }
-    }
-
-    @Test
-    fun given_the_user_provides_partial_consent_for_gdpr_then_user_consents_should_return_proper_consent() = runBlocking<Unit> {
-
-        val v7Consent = JSONObject(TestData.storedConsentV741)
-        val spClient = mockk<SpClient>(relaxed = true)
-
-        loadKoinModules(
-            mockModule(
-                spConfig = spConfGdpr,
-                gdprPmId = "488393",
-                spClientObserver = listOf(spClient),
-                diagnostic = v7Consent.toList()
-            )
-        )
-
-        wr {
-            val gdprAppliesFromUserConsents = userConsents(appContext).gdpr?.consent?.applies
-            gdprAppliesFromUserConsents.assertEquals(true)
         }
     }
